requirements.txt
pandas
openpyxl
python-docx
playwright
pillow
tqdm
qrcode
requests
